+++ sqlstore/migrations/0001_initital.sql
@@ -1,0 +1,54 @@
+-- +goose Up
+create table blocks
+(
+    vega_time     TIMESTAMP WITH TIME ZONE NOT NULL PRIMARY KEY,
+    height        BIGINT                   NOT NULL,
+    hash          BYTEA                    NOT NULL
+);
+
+create table assets
+(
+    id             BYTEA NOT NULL PRIMARY KEY,
+    name           TEXT NOT NULL UNIQUE,
+    symbol         TEXT NOT NULL UNIQUE,
+    total_supply   NUMERIC(32, 0),
+    decimals       INT,
+    quantum        INT,
+    source         TEXT,
+    erc20_contract TEXT
+);
+
+create table parties
+(
+    id        BYTEA NOT NULL PRIMARY KEY,
+    vega_time TIMESTAMP WITH TIME ZONE REFERENCES blocks (vega_time)
+);
+
+create table accounts
+(
+    id        SERIAL PRIMARY KEY,
+    party_id  BYTEA                    NOT NULL REFERENCES parties (id),
+    asset_id  BYTEA                    NOT NULL REFERENCES assets (id),
+    market_id BYTEA,
+    type      INT,
+    vega_time TIMESTAMP WITH TIME ZONE NOT NULL REFERENCES blocks(vega_time)
+);
+
+create table ledger
+(
+    id              BYTEA                    NOT NULL PRIMARY KEY,
+    account_from_id INT                      NOT NULL REFERENCES accounts(id),
+    account_to_id   INT                      NOT NULL REFERENCES accounts(id),
+    quantity        NUMERIC(32, 0)           NOT NULL,
+    vega_time       TIMESTAMP WITH TIME ZONE NOT NULL REFERENCES blocks(vega_time),
+    transfer_time   TIMESTAMP WITH TIME ZONE NOT NULL,
+    reference       TEXT,
+    type            TEXT
+);
+
+-- +goose Down
+drop table ledger;
+drop table accounts;
+drop table parties;
+drop table assets;
+drop table blocks;