+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - [296](https://github.com/vegaprotocol/data-node/pull/296) - Expose validator performance score attributes on Node object 
 - [298](https://github.com/vegaprotocol/data-node/pull/298) - Remove creation of vendor directory
 - [304](https://github.com/vegaprotocol/data-node/pull/304) - Added endpoint to support multiple versions of transaction request
+- [316](https://github.com/vegaprotocol/data-node/pull/316) - Add basic framework for connecting to `postgres` database
 
 ### 🐛 Fixes
 - [277](https://github.com/vegaprotocol/data-node/pull/277) - Now returns not-found error instead of internal error when proposal not found